refactor: extract shared CLI MainDeps helper (task-023 REFACTOR)
The `MainDeps` interface and the default `cwd/exit/log/errLog`
resolution logic was duplicated byte-for-byte between
`src/build-skills.ts` (from task 008) and the new
`src/skills-guard.ts` (from task 023). Extract both into
`src/cli-helpers.ts` so future CLI entry points can reuse them.

- New `src/cli-helpers.ts` exports `MainDeps`, `ResolvedMainDeps`,
  and `resolveMainDeps()`.
- `build-skills.ts` and `skills-guard.ts` both re-export `MainDeps`
  from the shared module (preserving their existing public surface)
  and use `resolveMainDeps()` in their `main()` bodies.

Behavior is unchanged; all 442 existing tests continue to pass.
